Consume single SparkBarSeries in SparkBars
