+++ packages/polaris-viz-core/src/components/index.ts
@@ -3,3 +3,4 @@ export {LinearGradientWithStops} from './LinearGradientWithStops';
 export {LineSeries} from './LineSeries';
 export {Bar} from './Bar';
 export type {PolarisVizProviderProps} from './PolarisVizProvider';
+export {SparkBarSeries} from './SparkBarSeries';
