+++ packages/polaris-viz/src/components/SparkBarChart/Chart.tsx
@@ -1,19 +1,10 @@
-import React, {useMemo} from 'react';
+import React from 'react';
 import {useTransition} from '@react-spring/web';
 import {
-  LinearGradientWithStops,
-  getSeriesColors,
-  useTheme,
-  Bar,
-  useSparkBar,
   Dimensions,
   SparkBarChartProps,
-  BARS_TRANSITION_CONFIG,
-  getAnimationTrail,
   ANIMATION_MARGIN,
-  STROKE_WIDTH,
-  uniqueId,
-  BORDER_RADIUS,
+  SparkBarSeries,
 } from '@shopify/polaris-viz-core';
 
 import {usePrefersReducedMotion} from '../../hooks';
@@ -35,46 +26,9 @@ export function Chart({
   theme,
 }: Props) {
   const {prefersReducedMotion} = usePrefersReducedMotion();
-  const selectedTheme = useTheme(theme);
-  const [seriesColor] = getSeriesColors(1, selectedTheme);
-
-  const {width, height} = dimensions ?? {width: 0, height: 0};
-  const id = useMemo(() => uniqueId('sparkbar'), []);
-  const clipId = useMemo(() => uniqueId('clip'), []);
-
   const shouldAnimate = !prefersReducedMotion && isAnimated;
 
-  const {
-    borderRadius,
-    dataWithIndex,
-    color,
-    getBarHeight,
-    strokeDasharray,
-    strokeDashoffset,
-    lineShape,
-    comparisonData,
-    xScale,
-    yScale,
-    barWidth,
-  } = useSparkBar({
-    data,
-    height,
-    dataOffsetLeft,
-    dataOffsetRight,
-    width,
-    seriesColor,
-  });
-
-  const transitions = useTransition(dataWithIndex, {
-    key: ({index}: {index: number}) => index,
-    from: {height: 0},
-    leave: {height: 0},
-    enter: ({value}) => ({height: getBarHeight(value == null ? 0 : value)}),
-    update: ({value}) => ({height: getBarHeight(value == null ? 0 : value)}),
-    default: {immediate: !shouldAnimate},
-    trail: shouldAnimate ? getAnimationTrail(dataWithIndex.length) : 0,
-    config: BARS_TRANSITION_CONFIG,
-  });
+  const {width, height} = dimensions ?? {width: 0, height: 0};
 
   const viewboxHeight = height + ANIMATION_MARGIN * 2;
 
@@ -95,62 +49,16 @@ export function Chart({
         height={viewboxHeight}
         width={width}
       >
-        <defs>
-          <LinearGradientWithStops
-            id={id}
-            gradient={color}
-            gradientUnits="userSpaceOnUse"
-            y1="100%"
-            y2="0%"
-          />
-        </defs>
-
-        <mask id={clipId}>
-          <g opacity={comparisonData ? '0.9' : '1'}>
-            {transitions(({height: barHeight}, item, _transition, index) => {
-              const xPosition = xScale(index.toString());
-              const height = shouldAnimate
-                ? barHeight
-                : getBarHeight(item.value ?? 0);
-
-              return (
-                <Bar
-                  borderRadius={
-                    selectedTheme.bar.hasRoundedCorners
-                      ? borderRadius
-                      : BORDER_RADIUS.None
-                  }
-                  key={index}
-                  x={xPosition == null ? 0 : xPosition}
-                  yScale={yScale}
-                  value={item.value.value}
-                  width={barWidth}
-                  height={height}
-                  fill="white"
-                />
-              );
-            })}
-          </g>
-        </mask>
-
-        <rect
-          fill={`url(#${id})`}
-          width={width}
+        <SparkBarSeries
+          data={data}
+          dataOffsetLeft={dataOffsetLeft}
+          dataOffsetRight={dataOffsetRight}
           height={height}
-          mask={`url(#${clipId})`}
+          shouldAnimate={shouldAnimate}
+          theme={theme}
+          useTransition={useTransition}
+          width={width}
         />
-
-        {comparisonData == null ? null : (
-          <path
-            stroke={selectedTheme.seriesColors.comparison}
-            strokeWidth={STROKE_WIDTH}
-            d={lineShape!}
-            className={styles.ComparisonLine}
-            opacity="0.9"
-            strokeDashoffset={strokeDashoffset}
-            strokeDasharray={strokeDasharray}
-          />
-        )}
       </svg>
     </React.Fragment>
   );
